% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{befftafall2018}
\alias{befftafall2018}
\alias{befftafall2018-package}
\title{befftafall2018: becoming an effective technology analyst.}
\description{
This package hosts assignment material for the 'becoming an effective technology analyst'
  fall 2018 class as part of the Strathclyde Business School MSc fintech program coursework.
}
\section{assignment}{

Instruction for the assigment.
}

\section{solutions}{

Example solutions to the assignment.
}

